Fix async-flow bugs; process input validations in series (but still asynchronously) so that evaluation order is predictable.
Ensuring that lists of fields and input values are evaluated in order makes
debugging easier, but also revealed a large number of async-processing bugs that
were leading to hangs.

Currently, I've judged sanity (and ease of debugging) more important than any
theoretical I/O parallelism; we may want to provide execution-order
guarantees to callers in other modules and/or user-provided validation rules or
conditions.

It's easy to switch from `eachSeries` back to `each` if we change our mind for
some reason, but if we end up wanting to, let's do it much closer to our initial
release.