Remove now-unnecessary brfs module.
I'm admittedly coming around to the fact that this was a good solution and made
a lot of sense, but the current tree is doing the file inclusion/rewriting all
on its own, and it saves us a browserify argument.

Apologies to Charles; I should have done a bit more review before jumping in and
writing code. Live and learn, I guess.