Allow fields to use self-referential structured validation.
For Gareth: I'm removing this limitation, as the case it's rejecting seems
totally legitimate (albeit a little weird) to refer to use a structured
validation to refer to the field that the structured validation is contained
within. One of our test cases (actually a couple) do this, and it seemed more
sensible than modifying the tests. Ping me if I'm totally wrong on this.